+++ frontend/src/comp/user/Form.jsx
@@ -1,0 +1,75 @@
+import React, { useState } from 'react';
+
+const Form = () => {
+  const [formData, setFormData] = useState({
+    name: '',
+    email: '',
+    password: '',
+  });
+
+  const handleChange = (e) => {
+    setFormData({
+      ...formData,
+      [e.target.name]: e.target.value,
+    });
+  };
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    console.log('Form submitted:', formData);
+    alert('Account created!');
+  };
+
+  return (
+    <div className='max-w-md mx-auto mt-8 p-6 bg-white rounded-lg shadow-md'>
+      <h2 className='text-2xl font-bold mb-6 text-center'>Sign Up</h2>
+
+      <form onSubmit={handleSubmit} className='space-y-4'>
+        <div>
+          <label className='block text-sm font-medium mb-1'>Name</label>
+          <input
+            type='text'
+            name='name'
+            value={formData.name}
+            onChange={handleChange}
+            className='w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500'
+            required
+          />
+        </div>
+
+        <div>
+          <label className='block text-sm font-medium mb-1'>Email</label>
+          <input
+            type='email'
+            name='email'
+            value={formData.email}
+            onChange={handleChange}
+            className='w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500'
+            required
+          />
+        </div>
+
+        <div>
+          <label className='block text-sm font-medium mb-1'>Password</label>
+          <input
+            type='password'
+            name='password'
+            value={formData.password}
+            onChange={handleChange}
+            className='w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500'
+            required
+          />
+        </div>
+
+        <button
+          type='submit'
+          className='w-full bg-blue-600 text-white py-2 px-4 rounded-md hover:bg-blue-700 transition duration-200'
+        >
+          Sign Up
+        </button>
+      </form>
+    </div>
+  );
+};
+
+export default Form;
