+++ frontend/src/pages/Admin.jsx
@@ -5,9 +5,7 @@ export const Admin = () => {
     <>
       <div className="p-6 text-center">
         <h1 className="text-3xl font-bold">Admin Page</h1>
-        <p>Welcome XX, here are your subscriptions.</p>
-        <button>Add Subscription</button>
-
+        <p>Welcome XX, here are your subscriptions.</p> {/* Placeholder for admin name */}
         <button>Log out</button>
       </div>
     </>
